# backend/src/controllers/page.controller.ts
import {
  Controller,
  Get,
  Post,
  Patch,
  Delete,
  Body,
  Param,
  Request,
  UseGuards,
  HttpCode,
  HttpStatus,
} from '@nestjs/common';
import { PageService } from '../services/page.service';
import { UserRepository } from '../repositories/user.repository';
import { FavoriteRepository } from '../repositories/favorite.repository';
import { CreatePageDto, UpdatePageDto } from '../dto/page.dto';
import { JwtAuthGuard } from '../guards/jwt-auth.guard';

@Controller('workspaces/:workspaceUuid/pages')
@UseGuards(JwtAuthGuard)
export class PageController {
  constructor(
    private readonly pageService: PageService,
    private readonly userRepo: UserRepository,
    private readonly favoriteRepo: FavoriteRepository
  ) {}

  @Post()
  @HttpCode(HttpStatus.CREATED)
  async createPage(
    @Param('workspaceUuid') workspaceUuid: string,
    @Request() req: any,
    @Body() dto: CreatePageDto
  ) {
    const userUuid = req.user.userId;
    const user = await this.userRepo.findByUuid(userUuid);
    if (!user) {
      throw new Error('User not found');
    }

    const page = await this.pageService.createPage(workspaceUuid, user.id, dto);
    return {
      success: true,
      data: page,
    };
  }

  @Get()
  async getPages(@Param('workspaceUuid') workspaceUuid: string, @Request() req: any) {
    const userUuid = req.user.userId;
    const user = await this.userRepo.findByUuid(userUuid);
    if (!user) {
      throw new Error('User not found');
    }

    const pages = await this.pageService.getPagesByWorkspace(workspaceUuid, user.id);
    return {
      success: true,
      data: pages,
    };
  }

  @Get('favorites')
  async getFavorites(@Request() req: any) {
    const userUuid = req.user.userId;
    const user = await this.userRepo.findByUuid(userUuid);
    if (!user) {
      throw new Error('User not found');
    }

    const favorites = await this.favoriteRepo.findByUserId(user.id);
    const pages = favorites.map((fav) => ({
      uuid: fav.page.uuid,
      title: fav.page.title,
      icon: fav.page.icon,
      type: fav.page.type,
      visibility: fav.page.visibility,
      updated_at: fav.page.updatedAt,
      favorited_at: fav.createdAt,
    }));

    return {
      success: true,
      data: pages,
    };
  }

  @Get(':pageUuid')
  async getPage(@Param('pageUuid') pageUuid: string, @Request() req: any) {
    const userUuid = req.user.userId;
    const user = await this.userRepo.findByUuid(userUuid);
    if (!user) {
      throw new Error('User not found');
    }

    const page = await this.pageService.getPage(pageUuid, user.id);
    return {
      success: true,
      data: page,
    };
  }

  @Patch(':pageUuid')
  async updatePage(
    @Param('pageUuid') pageUuid: string,
    @Request() req: any,
    @Body() dto: UpdatePageDto
  ) {
    const userUuid = req.user.userId;
    const user = await this.userRepo.findByUuid(userUuid);
    if (!user) {
      throw new Error('User not found');
    }

    const page = await this.pageService.updatePage(pageUuid, user.id, dto);
    return {
      success: true,
      data: page,
    };
  }

  @Delete(':pageUuid')
  @HttpCode(HttpStatus.NO_CONTENT)
  async deletePage(@Param('pageUuid') pageUuid: string, @Request() req: any) {
    const userUuid = req.user.userId;
    const user = await this.userRepo.findByUuid(userUuid);
    if (!user) {
      throw new Error('User not found');
    }

    await this.pageService.deletePage(pageUuid, user.id);
  }

  @Post(':pageUuid/favorite')
  @HttpCode(HttpStatus.CREATED)
  async addFavorite(@Param('pageUuid') pageUuid: string, @Request() req: any) {
    const userUuid = req.user.userId;
    const user = await this.userRepo.findByUuid(userUuid);
    if (!user) {
      throw new Error('User not found');
    }

    const page = await this.pageService.getPage(pageUuid, user.id);
    if (!page) {
      throw new Error('Page not found');
    }

    // Check if already favorited
    const existing = await this.favoriteRepo.findByUserAndPage(user.id, page.id);
    if (existing) {
      throw new Error('Page is already in favorites');
    }

    const favorite = await this.favoriteRepo.create({
      userId: user.id,
      pageId: page.id,
    });

    return {
      success: true,
      data: {
        page_id: pageUuid,
        created_at: favorite.createdAt,
      },
    };
  }

  @Delete(':pageUuid/favorite')
  @HttpCode(HttpStatus.NO_CONTENT)
  async removeFavorite(@Param('pageUuid') pageUuid: string, @Request() req: any) {
    const userUuid = req.user.userId;
    const user = await this.userRepo.findByUuid(userUuid);
    if (!user) {
      throw new Error('User not found');
    }

    const page = await this.pageService.getPage(pageUuid, user.id);
    if (!page) {
      throw new Error('Page not found');
    }

    const favorite = await this.favoriteRepo.findByUserAndPage(user.id, page.id);
    if (!favorite) {
      throw new Error('Favorite not found');
    }

    await this.favoriteRepo.deleteByUserIdAndPageId(user.id, page.id);
  }
}