# specs/001-technical-spec/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: All-in-One Workspace Platform API
  description: RESTful API for All-in-One Workspace Platform (Notion-like)
  version: 1.0.0
  contact:
    name: API Support
    email: api@workspace-platform.com
servers:
  - url: https://api.workspace-platform.com/api/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Workspaces
...
  /workspaces/{workspaceUuid}/pages:
    get:
      tags: [Pages]
      summary: List pages in workspace
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceUuid'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: parent_id
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            type: string
            enum: [page, database]
      responses:
        '200':
          description: List of pages
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Page'

    post:
      tags: [Pages]
      summary: Create page
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceUuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                parent_id:
                  type: string
                  format: uuid
                type:
                  type: string
                  enum: [page, database]
                  default: page
                icon:
                  type: string
                visibility:
                  type: string
                  enum: [private, workspace, public]
                  default: workspace
      responses:
        '201':
          description: Page created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
  /workspaces/{workspaceUuid}/pages/{pageUuid}:
    get:
      tags: [Pages]
      summary: Get page
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceUuid'
        - $ref: '#/components/parameters/PageUuid'
      responses:
        '200':
          description: Page details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'

    patch:
      tags: [Pages]
      summary: Update page
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceUuid'
        - $ref: '#/components/parameters/PageUuid'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                icon:
                  type: string
                cover_image:
                  type: string
                visibility:
                  type: string
                  enum: [private, workspace, public]
      responses:
        '200':
          description: Page updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'

    delete:
      tags: [Pages]
      summary: Delete page
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceUuid'
        - $ref: '#/components/parameters/PageUuid'
      responses:
        '204':
          description: Page deleted

  # Blocks
...
  /workspaces/{workspaceUuid}/pages/{pageUuid}/favorite:
    post:
      tags: [Pages]
      summary: Add page to favorites
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceUuid'
        - $ref: '#/components/parameters/PageUuid'
      responses:
        '201':
          description: Page added to favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      page_id:
                        type: string
                        format: uuid
                      created_at:
                        type: string
                        format: date-time
        '409':
          $ref: '#/components/responses/Conflict'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Pages]
      summary: Remove page from favorites
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceUuid'
        - $ref: '#/components/parameters/PageUuid'
      responses:
        '204':
          description: Page removed from favorites
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceUuid}/pages/favorites:
    get:
      tags: [Pages]
      summary: List favorite pages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceUuid'
      responses:
        '200':
          description: List of favorite pages
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Page'

  # Databases
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceUuid:
      name: workspaceUuid
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageUuid:
      name: pageUuid
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  schemas:
    User:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatar_url:
          type: string
          nullable: true
        timezone:
          type: string
        locale:
          type: string
        created_at:
          type: string
          format: date-time
...
    Page:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        title:
          type: string
        icon:
          type: string
          nullable: true
        type:
          type: string
          enum: [page, database]
        visibility:
          type: string
          enum: [private, workspace, public]
        parent_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            tokens:
              $ref: '#/components/schemas/TokenPair'
    TokenRefreshResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            tokens:
              $ref: '#/components/schemas/TokenPair'
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        requestId:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                - field: email
                  message: Invalid email format
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'